Implement single job to delete all project content
