+++ app/modules/uploads/helpers/projectStorage.ts
@@ -1,0 +1,3 @@
+export function getProjectStorageDir(projectId: string) {
+  return `storage/${projectId}`;
+}
