+++ workers/runners/general.ts
@@ -2,21 +2,15 @@ import "app/modules/documents/documents";
 import "app/modules/storage/storage";
 import type { Job } from "bullmq";
 import deleteProject from "../general/deleteProject";
-import deleteProjectFiles from "../general/deleteProjectFiles";
-import deleteProjectRuns from "../general/deleteProjectRuns";
-import deleteProjectSessions from "../general/deleteProjectSessions";
+import deleteProjectData from "../general/deleteProjectData";
 import removeExpiredTeamAssignment from "../general/removeExpiredTeamAssignment";
 import removeFeatureFlagFromUsers from "../general/removeFeatureFlagFromUsers";
 
 export default async (job: Job) => {
   try {
     switch (job.name) {
-      case 'DELETE_PROJECT:FILES':
-        return deleteProjectFiles(job);
-      case 'DELETE_PROJECT:SESSIONS':
-        return deleteProjectSessions(job);
-      case 'DELETE_PROJECT:RUNS': {
-        return deleteProjectRuns(job);
+      case 'DELETE_PROJECT:DATA': {
+        return deleteProjectData(job);
       }
       case 'DELETE_PROJECT:FINISH': {
         return deleteProject(job);
